Add Group/PaymentAccountSettings controller with test
